# group4_intelligence.py
"""
group4_intelligence.py
======================
Layer 4 — Strategic Intelligence Layer

Higher-order strategic engines that sit above clinical-operations tooling:
intelligence updates, research-opportunity scoring, precision hepatology,
and future methodology horizon-scanning.
"""

from fastapi import APIRouter

from shared_models import (
    IntelligenceUpdateRequest,
    IntelligenceUpdateResponse,
    ResearchOpportunityRequest,
    ResearchOpportunityResponse,
    PrecisionHepatologyRequest,
    PrecisionHepatologyResponse,
    FutureMethodologyRequest,
    FutureMethodologyResponse,
)

router = APIRouter()


# =========================
# Hepatology Intelligence Update Engine
# =========================

@router.post(
    "/orchestrator/hepatology-intelligence-update",
    response_model=IntelligenceUpdateResponse,
    tags=["Hepatology Research OS"],
)
def hepatology_intelligence_update(req: IntelligenceUpdateRequest):

    disease = req.disease.lower()

    if disease == "hbv":
        return IntelligenceUpdateResponse(
            disease=req.disease,
            update_domain=req.update_domain,
            intelligence_sources=[
                "PubMed",
                "ClinicalTrials.gov",
                "AASLD",
                "EASL",
                "APASL",
                "International HBV Meeting",
                "ICE-HBV",
            ],
            new_biomarkers=["HBV RNA", "HBcrAg", "Immune Profiling"],
            new_therapeutics=[
                "siRNA",
                "ASO",
                "Capsid Assembly Modulators",
                "Therapeutic Vaccines",
            ],
            new_trial_designs=[
                "Biomarker-guided Stopping",
                "Adaptive Enrichment",
            ],
            new_statistical_methods=[
                "Joint Models",
                "Dynamic Treatment Regimes",
            ],
            publication_relevance=[
                "Functional Cure",
                "Biomarker-guided Withdrawal",
            ],
            potential_impact=[
                "May redefine stopping criteria",
                "May improve functional cure prediction",
            ],
        )

    return IntelligenceUpdateResponse(
        disease=req.disease,
        update_domain=req.update_domain,
        intelligence_sources=[],
        new_biomarkers=[],
        new_therapeutics=[],
        new_trial_designs=[],
        new_statistical_methods=[],
        publication_relevance=[],
        potential_impact=[],
    )


# =========================
# Research Opportunity Engine
# =========================

@router.post(
    "/orchestrator/research-opportunity",
    response_model=ResearchOpportunityResponse,
    tags=["Hepatology Research OS"],
)
def research_opportunity(req: ResearchOpportunityRequest):

    disease = req.disease.lower()

    if disease == "hbv":
        return ResearchOpportunityResponse(
            disease=req.disease,
            evidence_gap=[
                "HBcrAg-guided stopping lacks definitive randomized "
                "validation",
                "Long-term finite therapy outcomes remain uncertain",
            ],
            clinical_gap=[
                "Need for safer finite therapy strategies",
                "Improved relapse prediction after NA withdrawal",
            ],
            biomarker_gap=[
                "Combined HBV RNA + HBcrAg algorithms",
                "Immune signature validation",
            ],
            therapeutic_gap=[
                "Combination cure regimens",
                "siRNA-based finite therapy strategies",
            ],
            trial_design_gap=[
                "Adaptive biomarker-guided stopping trials",
                "Platform cure trials",
            ],
            publication_opportunities=[
                "Functional Cure",
                "Biomarker-guided Withdrawal",
                "Combination Cure Strategies",
            ],
            competitive_density="high",
            strategic_priority_score=92,
            recommended_next_studies=[
                "HBcrAg-guided stopping RCT",
                "AI-assisted relapse prediction study",
                "Combination cure platform trial",
            ],
        )

    elif disease in ["masld", "mash"]:
        return ResearchOpportunityResponse(
            disease=req.disease,
            evidence_gap=[
                "Histology-free endpoint validation remains incomplete"
            ],
            clinical_gap=["Identification of rapid fibrosis progressors"],
            biomarker_gap=["ELF and PRO-C3 integration"],
            therapeutic_gap=["Combination therapy sequencing"],
            trial_design_gap=["Adaptive enrichment trials"],
            publication_opportunities=[
                "Histology-free Endpoints",
                "Precision Fibrosis Enrichment",
            ],
            competitive_density="very high",
            strategic_priority_score=88,
            recommended_next_studies=[
                "Histology-free endpoint validation study",
                "Precision enrichment trial",
            ],
        )

    elif disease == "hcc":
        return ResearchOpportunityResponse(
            disease=req.disease,
            evidence_gap=[
                "ctDNA-guided adjuvant strategies remain underdeveloped"
            ],
            clinical_gap=["Recurrence prediction after curative therapy"],
            biomarker_gap=["MRD validation"],
            therapeutic_gap=["Immunotherapy sequencing"],
            trial_design_gap=["Platform immunotherapy trials"],
            publication_opportunities=[
                "ctDNA-guided Therapy",
                "Platform Trial Design",
            ],
            competitive_density="very high",
            strategic_priority_score=95,
            recommended_next_studies=[
                "ctDNA-guided adjuvant RCT",
                "Adaptive platform immunotherapy trial",
            ],
        )

    return ResearchOpportunityResponse(
        disease=req.disease,
        evidence_gap=[],
        clinical_gap=[],
        biomarker_gap=[],
        therapeutic_gap=[],
        trial_design_gap=[],
        publication_opportunities=[],
        competitive_density="unknown",
        strategic_priority_score=0,
        recommended_next_studies=[],
    )


# =========================
# Precision Hepatology Engine
# =========================

@router.post(
    "/orchestrator/precision-hepatology",
    response_model=PrecisionHepatologyResponse,
    tags=["Hepatology Research OS"],
)
def precision_hepatology(req: PrecisionHepatologyRequest):

    disease = req.disease.lower()

    if disease == "hbv":
        return PrecisionHepatologyResponse(
            disease=req.disease,
            enrichment_strategies=[
                "HBcrAg-guided Enrichment",
                "HBV RNA-guided Enrichment",
            ],
            biomarker_strategies=["HBcrAg", "HBV RNA", "qHBsAg"],
            genetic_strategies=["Host Genetic Risk Profiling"],
            immune_strategies=[
                "T-cell Profiling",
                "Immune Signature Classification",
            ],
            ctDNA_strategies=[],
            risk_prediction_strategies=[
                "Relapse Prediction Models",
                "AI-assisted Risk Prediction",
            ],
            precision_trial_concepts=[
                "Biomarker-guided Stopping Trial",
                "Precision Functional Cure Trial",
            ],
        )

    elif disease in ["masld", "mash"]:
        return PrecisionHepatologyResponse(
            disease=req.disease,
            enrichment_strategies=[
                "Fibrosis Enrichment",
                "High-risk Metabolic Phenotype",
            ],
            biomarker_strategies=["ELF", "PRO-C3"],
            genetic_strategies=["Polygenic Risk Scores"],
            immune_strategies=[],
            ctDNA_strategies=[],
            risk_prediction_strategies=["Fibrosis Progression Models"],
            precision_trial_concepts=[
                "Precision Fibrosis Trial",
                "Biomarker-enriched MASH Trial",
            ],
        )

    elif disease == "hcc":
        return PrecisionHepatologyResponse(
            disease=req.disease,
            enrichment_strategies=[
                "ctDNA-positive Population",
                "MRD-positive Population",
            ],
            biomarker_strategies=["AFP", "ctDNA", "MRD"],
            genetic_strategies=["Molecular Subtype Selection"],
            immune_strategies=["Immune Microenvironment Profiling"],
            ctDNA_strategies=["ctDNA-guided Adjuvant Therapy"],
            risk_prediction_strategies=["Recurrence Prediction Models"],
            precision_trial_concepts=[
                "MRD-guided Adjuvant Trial",
                "ctDNA-guided Surveillance Trial",
            ],
        )

    return PrecisionHepatologyResponse(
        disease=req.disease,
        enrichment_strategies=[],
        biomarker_strategies=[],
        genetic_strategies=[],
        immune_strategies=[],
        ctDNA_strategies=[],
        risk_prediction_strategies=[],
        precision_trial_concepts=[],
    )


# =========================
# Future Methodology Engine
# =========================

@router.post(
    "/orchestrator/future-methodology",
    response_model=FutureMethodologyResponse,
    tags=["Hepatology Research OS"],
)
def future_methodology(req: FutureMethodologyRequest):

    disease = req.disease.lower()

    if disease == "hbv":
        return FutureMethodologyResponse(
            disease=req.disease,
            emerging_trial_designs=[
                "Adaptive Platform Trials",
                "Biomarker-guided Adaptive Trials",
                "Master Protocols",
            ],
            emerging_statistical_methods=[
                "Bayesian Learning Systems",
                "Dynamic Treatment Regimes",
                "Joint Models",
            ],
            digital_biomarker_trends=[
                "Continuous Biomarker Monitoring",
                "AI-derived Biomarkers",
            ],
            ai_trial_design_trends=[
                "AI-assisted Eligibility Selection",
                "AI-guided Adaptive Design",
            ],
            synthetic_control_trends=[
                "External Control Arms",
                "Synthetic Cohorts",
            ],
            federated_trial_trends=["Multi-center Federated Networks"],
            adoption_timeline=[
                "2025-2030: Biomarker-guided Trials",
                "2030-2035: AI-assisted Trial Design",
                "2035-2045: Learning Trial Systems",
            ],
            future_impact_assessment=[
                "May redefine finite therapy studies",
                "May accelerate functional cure development",
            ],
        )

    elif disease in ["masld", "mash"]:
        return FutureMethodologyResponse(
            disease=req.disease,
            emerging_trial_designs=[
                "Adaptive Enrichment",
                "Master Protocols",
            ],
            emerging_statistical_methods=[
                "Bayesian Adaptive Design",
                "Joint Models",
            ],
            digital_biomarker_trends=["Digital Histology", "AI Pathology"],
            ai_trial_design_trends=["AI-powered Patient Enrichment"],
            synthetic_control_trends=["Synthetic Comparator Arms"],
            federated_trial_trends=["Global MASLD Research Networks"],
            adoption_timeline=[
                "2025-2030: Histology-free Endpoints",
                "2030-2035: AI-guided Enrichment",
                "2035-2045: Fully Adaptive Precision Trials",
            ],
            future_impact_assessment=[
                "May reduce biopsy dependence",
                "May improve trial efficiency",
            ],
        )

    elif disease == "hcc":
        return FutureMethodologyResponse(
            disease=req.disease,
            emerging_trial_designs=[
                "Platform Immunotherapy Trials",
                "Adaptive Platform Trials",
            ],
            emerging_statistical_methods=[
                "Joint Frailty Models",
                "Multi-state Models",
            ],
            digital_biomarker_trends=[
                "ctDNA Monitoring",
                "MRD-guided Therapy",
            ],
            ai_trial_design_trends=["AI-guided Recurrence Prediction"],
            synthetic_control_trends=["External Control Integration"],
            federated_trial_trends=["Global HCC Trial Networks"],
            adoption_timeline=[
                "2025-2030: ctDNA-guided Trials",
                "2030-2035: Adaptive Platform Expansion",
                "2035-2045: Precision Oncology Ecosystems",
            ],
            future_impact_assessment=[
                "May transform adjuvant therapy strategies",
                "May personalize recurrence prevention",
            ],
        )

    return FutureMethodologyResponse(
        disease=req.disease,
        emerging_trial_designs=[],
        emerging_statistical_methods=[],
        digital_biomarker_trends=[],
        ai_trial_design_trends=[],
        synthetic_control_trends=[],
        federated_trial_trends=[],
        adoption_timeline=[],
        future_impact_assessment=[],
    )