# openapi_gpt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hepatology Research OS (GPT Surface)",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://clinicaltrialapi-dqml.onrender.com"
    }
  ],
  "paths": {
...
    "/orchestrator/research-opportunity": {
      "post": {
        "tags": [
          "Hepatology Research OS"
        ],
        "summary": "Research Opportunity",
        "operationId": "research_opportunity_orchestrator_research_opportunity_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchOpportunityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchOpportunityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orchestrator/precision-hepatology": {
      "post": {
        "tags": [
          "Hepatology Research OS"
        ],
        "summary": "Precision Hepatology",
        "operationId": "precision_hepatology_orchestrator_precision_hepatology_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrecisionHepatologyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrecisionHepatologyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orchestrator/future-methodology": {
      "post": {
        "tags": [
          "Hepatology Research OS"
        ],
        "summary": "Future Methodology",
        "operationId": "future_methodology_orchestrator_future_methodology_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FutureMethodologyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FutureMethodologyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FutureMethodologyRequest": {
        "properties": {
          "disease": {
            "type": "string",
            "title": "Disease"
          }
        },
        "type": "object",
        "required": [
          "disease"
        ],
        "title": "FutureMethodologyRequest"
      },
      "FutureMethodologyResponse": {
        "properties": {
          "disease": {
            "type": "string",
            "title": "Disease"
          },
          "emerging_trial_designs": {
            "items": {},
            "type": "array",
            "title": "Emerging Trial Designs"
          },
          "emerging_statistical_methods": {
            "items": {},
            "type": "array",
            "title": "Emerging Statistical Methods"
          },
          "digital_biomarker_trends": {
            "items": {},
            "type": "array",
            "title": "Digital Biomarker Trends"
          },
          "ai_trial_design_trends": {
            "items": {},
            "type": "array",
            "title": "Ai Trial Design Trends"
          },
          "synthetic_control_trends": {
            "items": {},
            "type": "array",
            "title": "Synthetic Control Trends"
          },
          "federated_trial_trends": {
            "items": {},
            "type": "array",
            "title": "Federated Trial Trends"
          },
          "adoption_timeline": {
            "items": {},
            "type": "array",
            "title": "Adoption Timeline"
          },
          "future_impact_assessment": {
            "items": {},
            "type": "array",
            "title": "Future Impact Assessment"
          }
        },
        "type": "object",
        "required": [
          "disease",
          "emerging_trial_designs",
          "emerging_statistical_methods",
          "digital_biomarker_trends",
          "ai_trial_design_trends",
          "synthetic_control_trends",
          "federated_trial_trends",
          "adoption_timeline",
          "future_impact_assessment"
        ],
        "title": "FutureMethodologyResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrecisionHepatologyRequest": {
        "properties": {
          "disease": {
            "type": "string",
            "title": "Disease"
          }
        },
        "type": "object",
        "required": [
          "disease"
        ],
        "title": "PrecisionHepatologyRequest"
      },
      "PrecisionHepatologyResponse": {
        "properties": {
          "disease": {
            "type": "string",
            "title": "Disease"
          },
          "enrichment_strategies": {
            "items": {},
            "type": "array",
            "title": "Enrichment Strategies"
          },
          "biomarker_strategies": {
            "items": {},
            "type": "array",
            "title": "Biomarker Strategies"
          },
          "genetic_strategies": {
            "items": {},
            "type": "array",
            "title": "Genetic Strategies"
          },
          "immune_strategies": {
            "items": {},
            "type": "array",
            "title": "Immune Strategies"
          },
          "ctDNA_strategies": {
            "items": {},
            "type": "array",
            "title": "Ctdna Strategies"
          },
          "risk_prediction_strategies": {
            "items": {},
            "type": "array",
            "title": "Risk Prediction Strategies"
          },
          "precision_trial_concepts": {
            "items": {},
            "type": "array",
            "title": "Precision Trial Concepts"
          }
        },
        "type": "object",
        "required": [
          "disease",
          "enrichment_strategies",
          "biomarker_strategies",
          "genetic_strategies",
          "immune_strategies",
          "ctDNA_strategies",
          "risk_prediction_strategies",
          "precision_trial_concepts"
        ],
        "title": "PrecisionHepatologyResponse"
      },
...
      "ResearchOpportunityRequest": {
        "properties": {
          "disease": {
            "type": "string",
            "title": "Disease"
          },
          "research_focus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Research Focus"
          }
        },
        "type": "object",
        "required": [
          "disease"
        ],
        "title": "ResearchOpportunityRequest"
      },
      "ResearchOpportunityResponse": {
        "properties": {
          "disease": {
            "type": "string",
            "title": "Disease"
          },
          "evidence_gap": {
            "items": {},
            "type": "array",
            "title": "Evidence Gap"
          },
          "clinical_gap": {
            "items": {},
            "type": "array",
            "title": "Clinical Gap"
          },
          "biomarker_gap": {
            "items": {},
            "type": "array",
            "title": "Biomarker Gap"
          },
          "therapeutic_gap": {
            "items": {},
            "type": "array",
            "title": "Therapeutic Gap"
          },
          "trial_design_gap": {
            "items": {},
            "type": "array",
            "title": "Trial Design Gap"
          },
          "publication_opportunities": {
            "items": {},
            "type": "array",
            "title": "Publication Opportunities"
          },
          "competitive_density": {
            "type": "string",
            "title": "Competitive Density"
          },
          "strategic_priority_score": {
            "type": "integer",
            "title": "Strategic Priority Score"
          },
          "recommended_next_studies": {
            "items": {},
            "type": "array",
            "title": "Recommended Next Studies"
          }
        },
        "type": "object",
        "required": [
          "disease",
          "evidence_gap",
          "clinical_gap",
          "biomarker_gap",
          "therapeutic_gap",
          "trial_design_gap",
          "publication_opportunities",
          "competitive_density",
          "strategic_priority_score",
          "recommended_next_studies"
        ],
        "title": "ResearchOpportunityResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}